Implemented User Login and User Sign-Up on the first activity when app opens using Firebase Authentication implementation.
Next on the list of things to do will be to create a "sign out" button to return to the first activity once again when the user decided to sign out.

Once the "sign out" part is sorted, the questions written by the authenticated user will be stored to only be accessed by that specific user. In other words, the user can write questions to their part of the database and retrieve questions from their part of the database associated with their UID.